refactor: extract pipeline/persistence.rs into astra-evolution (1270 LOC)
Moves PersistedActiveCanary into astra-evolution::types and the full pipeline
persistence layer into astra-evolution::persistence. astra-evolution already
depends on astra-pipeline, so it is the natural home for the module that
bridges pipeline snapshots with evolution canary state.

- runtime/src/pipeline/persistence.rs is now a 6-line re-export stub
- astra-evolution gains a `dirs` dep for the home-dir helper
- No behavior changes; all paths resolve via re-exports
