+++ rust/crates/astra-evolution/Cargo.toml
@@ -8,6 +8,7 @@ description = "Multi-axis self-evolution engine types and core logic for Astra"
 serde = { workspace = true }
 serde_json = { workspace = true }
 tracing = { workspace = true }
+dirs = "6.0.0"
 
 # Internal crates
 astra-pipeline = { path = "../astra-pipeline" }
